Added different UIs for each inclusion type
Also removed the "Select..." option when choosing multiple criteria with the "Any" inclusion mode, because it used to count as a criteria in the $inclusion_criteria array.
Still need to fix the checkbox in the "All" mode. Will focus on improving UI as the last step. Next to do is adding the mapping table in the database.